fix(frontend): several position-related fixes
